Add files via upload
Files for DeadlineTimers project. Goal is for user to be able to create timers with a given end date that are displayed and tick down until that date. The idea for this was inspired by the various ways the game Frostpunk showed approaching quest/demand deadlines, and I created this project because I seemed to have no problem being motivated to complete tasks in Frostpunk yet struggled to stay motivated during school and when working on personal projects. 

The deprecated index file is the code for the project from sometime in the summer of 2024 up to October 30th of 2024. Code was overhauled on October 30th to use objects to store timers and with considerably better documentation. Bug fixes were also made. October 30-31 was primarily rewriting and bug fixing the newTimer() function and creating the displayTimer() function. November 1st involved bug-fixes and work focused on the tickTimers() function and subsequent related functions.