Update Index.js
Fixed logical error in tickTimers that would use the negativeTick if it was given a time that ended with a zero value.
Created calculateTickingTime() to store ticking time in timer object and so end times can be calculated when loading timers in is eventually added. 
Updated newTimer() to store calculated ticking time in timer.
Updated displayTimer to use tickingTime of given timer rather than generating a new ticking time.